backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
openpyxl>=3.1.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.6.0
joblib>=1.4.0
email-validator>=2.0.0
